+++ TodoList.Api/Program.cs
@@ -4,6 +4,7 @@ using TodoList.Infrastructure;
 using TodoList.Services.Areas;
 using TodoList.Services.Areas.Archive;
 using TodoList.Services.Areas.Items;
+using TodoList.Services.Areas.UnArchive;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -21,6 +22,8 @@ builder.Services.AddSwaggerGen();
 builder.Services.AddTransient<IItemService, ItemService>();
 
 builder.Services.AddTransient<IArchiveService, ArchiveService>();
+
+builder.Services.AddTransient<IUnArchiveService, UnArchiveService>();
 
 var app = builder.Build();
