+++ TodoList.Blazor/Pages/Archive/Items.razor
@@ -2,6 +2,7 @@
 
 @inject IJSRuntime Js
 @inject IArchiveService ArchiveService
+@inject IUnArchiveService UnArchiveService
 
 @if (_items == null)
 {
@@ -61,7 +62,7 @@ else
 
     private async Task UnArchiveByIdAsync(int id)
     {
-        await ArchiveService.UnArchiveByIdAsync(id);
+        await UnArchiveService.UnArchiveByIdAsync(id);
 
         _items = await ArchiveService.GetListAsync();
