+++ TodoList.Services/Areas/Archive/IArchiveService.cs
@@ -7,6 +7,5 @@ public interface IArchiveService
     public Task<IList<ArchivedDto>> GetListAsync();
     public Task<ArchivedDto> GetByIdAsync(int id);
     public Task<int> ArchiveByIdAsync(int id);
-    public Task<int> UnArchiveByIdAsync(int id);
     public Task DeleteByIdAsync(int id);
 }
