+++ TodoList.Services/Areas/UnArchive/IUnArchiveService.cs
@@ -1,0 +1,6 @@
+﻿namespace TodoList.Services.Areas.UnArchive;
+
+public interface IUnArchiveService
+{
+    public Task<int> UnArchiveByIdAsync(int id);
+}
